+++ evennia/contrib/base_systems/custom_gametime/custom_gametime.py
@@ -328,4 +328,4 @@ class GametimeScript(DefaultScript):
             callback()
 
         seconds = real_seconds_until(**self.db.gametime)
-        self.restart(interval=seconds)
+        self.start(interval=seconds, force_restart=True)