+++ evennia/prototypes/prototypes.py
@@ -528,10 +528,10 @@ def search_prototype(
 
     """
     # This will load the prototypes the first time they are searched
-    loaded = getattr(load_module_prototypes, '_LOADED', False)
+    loaded = getattr(load_module_prototypes, "_LOADED", False)
     if not loaded:
         load_module_prototypes()
-        setattr(load_module_prototypes, '_LOADED', True)
+        setattr(load_module_prototypes, "_LOADED", True)
 
     # prototype keys are always in lowecase
     if key: