+++ evennia/utils/dbserialize.py
@@ -18,19 +18,20 @@ in-situ, e.g `obj.db.mynestedlist[3][5] = 3` would never be saved and
 be out of sync with the database.
 
 """
+from collections import OrderedDict, defaultdict, deque
+from collections.abc import MutableMapping, MutableSequence, MutableSet
 from functools import update_wrapper
-from collections import deque, OrderedDict, defaultdict
-from collections.abc import MutableSequence, MutableSet, MutableMapping
 
 try:
-    from pickle import dumps, loads
+    from pickle import UnpicklingError, dumps, loads
 except ImportError:
     from pickle import dumps, loads
-from django.core.exceptions import ObjectDoesNotExist
+
 from django.contrib.contenttypes.models import ContentType
+from django.core.exceptions import ObjectDoesNotExist
 from django.utils.safestring import SafeString
-from evennia.utils.utils import uses_database, is_iter, to_bytes
 from evennia.utils import logger
+from evennia.utils.utils import is_iter, to_bytes, uses_database
 
 __all__ = ("to_pickle", "from_pickle", "do_pickle", "do_unpickle", "dbserialize", "dbunserialize")
 
@@ -238,6 +239,9 @@ class _SaverMutable(object):
 
     def __gt__(self, other):
         return self._data > other
+
+    def __or__(self, other):
+        return self._data | other
 
     @_save
     def __setitem__(self, key, value):
@@ -450,7 +454,9 @@ def deserialize(obj):
         elif tname in ("_SaverOrderedDict", "OrderedDict"):
             return OrderedDict([(_iter(key), _iter(val)) for key, val in obj.items()])
         elif tname in ("_SaverDefaultDict", "defaultdict"):
-            return defaultdict(obj.default_factory, {_iter(key): _iter(val) for key, val in obj.items()})
+            return defaultdict(
+                obj.default_factory, {_iter(key): _iter(val) for key, val in obj.items()}
+            )
         elif tname in _DESERIALIZE_MAPPING:
             return _DESERIALIZE_MAPPING[tname](_iter(val) for val in obj)
         elif is_iter(obj):
@@ -602,7 +608,9 @@ def to_pickle(data):
 
     def process_item(item):
         """Recursive processor and identification of data"""
+
         dtype = type(item)
+
         if dtype in (str, int, float, bool, bytes, SafeString):
             return item
         elif dtype == tuple:
@@ -612,7 +620,10 @@ def to_pickle(data):
         elif dtype in (dict, _SaverDict):
             return dict((process_item(key), process_item(val)) for key, val in item.items())
         elif dtype in (defaultdict, _SaverDefaultDict):
-            return defaultdict(item.default_factory, ((process_item(key), process_item(val)) for key, val in item.items()))
+            return defaultdict(
+                item.default_factory,
+                ((process_item(key), process_item(val)) for key, val in item.items()),
+            )
         elif dtype in (set, _SaverSet):
             return set(process_item(val) for val in item)
         elif dtype in (OrderedDict, _SaverOrderedDict):
@@ -620,7 +631,20 @@ def to_pickle(data):
         elif dtype in (deque, _SaverDeque):
             return deque(process_item(val) for val in item)
 
-        elif hasattr(item, "__iter__"):
+        # not one of the base types
+        if hasattr(item, "__serialize_dbobjs__"):
+            # Allows custom serialization of any dbobjects embedded in
+            # the item that Evennia will otherwise not find (these would
+            # otherwise lead to an error). Use the dbserialize helper from
+            # this method.
+            try:
+                item.__serialize_dbobjs__()
+            except TypeError as err:
+                # we catch typerrors so we can handle both classes (requiring
+                # classmethods) and instances
+                pass
+
+        if hasattr(item, "__iter__"):
             # we try to conserve the iterable class, if not convert to list
             try:
                 return item.__class__([process_item(val) for val in item])
@@ -678,7 +702,10 @@ def from_pickle(data, db_obj=None):
         elif dtype == dict:
             return dict((process_item(key), process_item(val)) for key, val in item.items())
         elif dtype == defaultdict:
-            return defaultdict(item.default_factory, ((process_item(key), process_item(val)) for key, val in item.items()))
+            return defaultdict(
+                item.default_factory,
+                ((process_item(key), process_item(val)) for key, val in item.items()),
+            )
         elif dtype == set:
             return set(process_item(val) for val in item)
         elif dtype == OrderedDict:
@@ -692,6 +719,22 @@ def from_pickle(data, db_obj=None):
                 return item.__class__(process_item(val) for val in item)
             except (AttributeError, TypeError):
                 return [process_item(val) for val in item]
+
+        if hasattr(item, "__deserialize_dbobjs__"):
+            # this allows the object to custom-deserialize any embedded dbobjs
+            # that we previously serialized with __serialize_dbobjs__.
+            # use the dbunserialize helper in this module.
+            try:
+                item.__deserialize_dbobjs__()
+            except (TypeError, UnpicklingError):
+                # handle recoveries both of classes (requiring classmethods
+                # or instances. Unpickling errors can happen when re-loading the
+                # data from cache (because the hidden entity was already
+                # deserialized and stored back on the object, unpickling it
+                # again fails). TODO: Maybe one could avoid this retry in a
+                # more graceful way?
+                pass
+
         return item
 
     def process_tree(item, parent):
@@ -744,6 +787,13 @@ def from_pickle(data, db_obj=None):
                 dat = _SaverList(_parent=parent)
                 dat._data.extend(process_tree(val, dat) for val in item)
                 return dat
+
+        if hasattr(item, "__deserialize_dbobjs__"):
+            try:
+                item.__deserialize_dbobjs__()
+            except (TypeError, UnpicklingError):
+                pass
+
         return item
 
     if db_obj: